op9: use fixed sizes for dynamic partitions
This is for development only.

We distribute images with super_empty, and the users will have a
fresh super image upon each installation.

During development, however, the partition sizes routinely
change and a fastboot flash on each partitions under fastbootd
will continue to cause logical fragmentation due to non-linear
mapping.

Non-linear mapping is kind of the whole point of using dynamic
partition, but we don't have to suffer the performance penalty
as we can just reserve a big size for each partition to keep the
mapping to be contiguous.

Change-Id: Ia984d1ab1f8ac70d416ba362d1a3fcf9edc32f26
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>